add zero

+++ clearml_training_wrapper.py
@@ -200,7 +200,7 @@ def create_training_params(custom_task_data_path, adapter_save_path):
         "lora_rank": args.lora_rank,
         "lora_alpha": args.lora_alpha,
         "lora_target": args.lora_target,
-        "deepspeed": "/app/config/ds_z3_config.json",
+        "deepspeed": '/app/config/ds_z' + str(args.ds_stage) + '_config.json',
         "flash_attn": "fa2",
 
         "dataset_dir": custom_task_data_path,
@@ -225,7 +225,6 @@ def create_training_params(custom_task_data_path, adapter_save_path):
         "warmup_ratio": 0.1,
         "bf16": True,
         "ddp_timeout": 180000000,
-
     }
 
     training_params_path = os.path.join(custom_task_data_path,'training_params.yaml')
@@ -260,10 +259,11 @@ if __name__ == '__main__':
     # General arguments
     parser.add_argument('--project_name', type=str, default='llm_factory_trainer', help='name of project')
     parser.add_argument('--task_name', type=str, default='trainer_template_v0', help='name of project')
+    parser.add_argument('--ds_stage', type=int, default=0, help='location of dataset')
     parser.add_argument('--lora_rank', type=int, default=8, help='location of dataset')
     parser.add_argument('--lora_alpha', type=int, default=8, help='location of dataset')
     parser.add_argument('--lora_target', type=str, default='all', help='location of dataset')
-    parser.add_argument('--batch_size', type=int, default=8, help='location of dataset')
+    parser.add_argument('--batch_size', type=int, default=1, help='location of dataset')
     parser.add_argument('--gradient_accumulation_steps', type=int, default=2, help='location of dataset')
     parser.add_argument('--epoch', type=float, default=1.0, help='location of dataset')
     parser.add_argument('--lr', type=float, default=1e-4, help='location of dataset')
@@ -305,11 +305,19 @@ if __name__ == '__main__':
     #os.environ["CLEARML_LOG_ENVIRONMENT"] = tmp_custom_task_data_path
     os.environ["CLEARML_TASK_NO_REUSE"] = '1'
     os.environ["CLEARML_LOG_LEVEL"] = 'INFO'
+    '''
 
     print('ENVS:')
     for name, value in os.environ.items():
         print("{0}: {1}".format(name, value))
-    '''
+
+    control_node = False
+
+    if "PMIX_RANK" in os.environ:
+        if os.environ["PMIX_RANK"] == '0':
+            control_node = True
+    else:
+        control_node = True
 
     print('Starting ClearML Task')
     task = Task.init(project_name=args.project_name, task_name=args.task_name)
@@ -342,7 +350,7 @@ if __name__ == '__main__':
         if rc != 0:
             raise ValueError(f"Training failed with return code {rc}")
 
-        if args.job_id is not None:
+        if (args.job_id is not None) and (control_node):
 
             # remove checkpoints
             print('Cleaning adapter before upload')
@@ -393,14 +401,14 @@ if __name__ == '__main__':
 
     #remove temp
     #shutil.rmtree(tmp_custom_task_data_path)
-
-    print('Finished Training, cleaning files')
-    clean_paths = [custom_task_data_path, dataset_cache_path]
-    for path in clean_paths:
-        if path is not None:
-            if os.path.exists(path):
-                print('Removing path:', path)
-                shutil.rmtree(path)
+    if control_node:
+        print('Finished Training, cleaning files')
+        clean_paths = [custom_task_data_path, dataset_cache_path]
+        for path in clean_paths:
+            if path is not None:
+                if os.path.exists(path):
+                    print('Removing path:', path)
+                    shutil.rmtree(path)
 
     #for name, value in os.environ.items():
     #    print("{0}: {1}".format(name, value))
